simplify utilities
